backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1

# RAG & LLM
langchain==0.3.14
langchain-community==0.3.14
langchain-chroma>=0.2.2
chromadb>=1.0.0

# Document Processing
PyMuPDF==1.25.3
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3

# Redis
redis==5.2.1

# Database
aiosqlite==0.22.1

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
